+++ app/code/community/Ffuenf/Common/Block/Adminhtml/Log/Abstract.php
@@ -24,47 +24,8 @@ abstract class Ffuenf_Common_Block_Adminhtml_Log_Abstract extends Mage_Adminhtml
     {
         parent::__construct();
         $this->_blockGroup = 'ffuenf_common';
+        $this->_controller = 'adminhtml_log_' . $this->_logType;
         $this->_removeButton('add');
-    }
-
-    protected function _prepareLayout()
-    {
-        if (null !== $this->_logType) {
-            $logFilePath = Ffuenf_Common_Model_Logger::getAbsoluteLogFilePath($this->_logType);
-            if (Mage::helper('ffuenf_common/file')->exists($logFilePath)) {
-                $this->_addButton(
-                    'download',
-                    array(
-                        'label'   => $this->_getDownloadButtonLabel(),
-                        'onclick' => 'setLocation(\'' . $this->_getDownloadUrl() . '\')',
-                        'class'   => 'scalable'
-                    ),
-                    -1
-                );
-            } else {
-                $this->_addButton(
-                    'download',
-                    array(
-                        'label'    => $this->_getDownloadButtonLabel(),
-                        'onclick'  => 'setLocation(\'' . $this->_getDownloadUrl() . '\')',
-                        'class'    => 'scalable',
-                        'disabled' => true
-                    ),
-                    -1
-                );
-            }
-        }
-        return parent::_prepareLayout();
-    }
-
-    protected function _getDownloadButtonLabel()
-    {
-        return $this->__('Download as CSV');
-    }
-
-    protected function _getDownloadUrl()
-    {
-        return $this->getUrl('*/*/download');
     }
 
     public function getHeaderCssClass()